third_party/WebKit/LayoutTests/security/immutable-prototype.html

<!DOCTYPE html>
<html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="immutable-prototype.js"></script>
<script>
'use strict';

test(() => {
  let result = prototypeChain(window);
  assert_array_equals(result,
                      ['immutable', 'immutable', 'immutable', 'immutable', 'immutable']);
}, 'global object prototype chain should be immutable');
</script>
</html>
